# trvCollection.postman_collection.json-Swagger20.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "trvCollection",
    "description": "# Introduction\nSimple express API with 5 main functionalities.\n * Managing rowing days appointments.\n * User authentification.\n * Managing news of the website.\n * Managing user data.\n * Enabling pdf download.\n",
    "contact": {}
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/appointments": {
      "get": {
        "description": "Get all reserved(saved in database) and available (empty default objects) rowing days appointments for specific days of the week(defined in table rowingdays)",
        "summary": "http://localhost:3000/api/appointments",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/appointments",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          }
        }
      },
      "post": {
        "description": "Reserve a rowing day appointment",
        "summary": "http://localhost:3000/api/appointments",
        "tags": [
          "Misc"
        ],
        "operationId": "Posthttp://localhost:3000/api/appointments",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1appointmentsrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "Please enter all fields!"
              }
            }
          }
        }
      },
      "put": {
        "description": "Edit a rowing day appointment",
        "summary": "http://localhost:3000/api/appointments",
        "tags": [
          "Misc"
        ],
        "operationId": "Puthttp://localhost:3000/api/appointments",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "required": true,
            "type": "integer",
            "format": "int32",
            "description": ""
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1appointmentsrequest1"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1appointments2"
            },
            "examples": {
              "application/json": {
                "successMsg": "Appointment created!"
              }
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "content-length": {
                "type": "string",
                "default": "37"
              },
              "etag": {
                "type": "string",
                "default": "W/\"25-kF/Ci1eTaztXrVLqFUsyJSVibVQ\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 19:50:11 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "Please enter all fields!"
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "No token, authorization denied!"
              }
            }
          }
        }
      }
    },
    "/rowingdays": {
      "put": {
        "description": "Change weekdays, time, amount and information of rowing days",
        "summary": "http://localhost:3000/api/rowingdays",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/rowingdays",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1rowingdaysrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "type": "object"
            },
            "headers": {}
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "No token, authorization denied!"
              }
            }
          }
        }
      }
    },
    "/auth": {
      "post": {
        "description": "Authentificate current user",
        "summary": "http://localhost:3000/api/auth",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/auth",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1authrequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1auth2"
            },
            "examples": {
              "application/json": {
                "token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6OSwicm9sZSI6IkFETUlOIiwiaWF0IjoxNjA3NDU3NDA4LCJleHAiOjE2MDc0NjQ2MDh9.LsbiY8yRg-zSSmZu8V4PaJRemxZ3w2LECXAbPw_28Ng",
                "user": {
                  "id": 9,
                  "name": "Tony Vercetti",
                  "email": "tony@mail.com",
                  "role": "ADMIN"
                }
              }
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "content-length": {
                "type": "string",
                "default": "247"
              },
              "etag": {
                "type": "string",
                "default": "W/\"f7-/UY/E99J3bq1MVrnXHcvjbyOBiQ\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 19:56:48 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1users"
            },
            "examples": {
              "application/json": {
                "msg": "invalid credentials!"
              }
            }
          }
        }
      }
    },
    "/auth/user": {
      "get": {
        "description": "Get a specific user",
        "summary": "http://localhost:3000/api/auth/user",
        "tags": [
          "Misc"
        ],
        "operationId": "http://localhost:3000/api/auth/user",
        "deprecated": false,
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "schema": {
              "$ref": "#/definitions/http%3A~1~1localhost%3A3000~1api~1auth~1user"
            },
            "examples": {
              "application/json": {
                "user": {
                  "id": 9,
                  "name": "Tony Vercetti",
                  "email": "tony@mail.com",
                  "role": "ADMIN"
                }
              }
            },
            "headers": {
              "x-powered-by": {
                "type": "string",
                "default": "Express"
              },
              "access-control-allow-origin": {
                "type": "string",
                "default": "*"
              },
              "content-type": {
                "type": "string",
                "default": "application/json; charset=utf-8"
              },
              "content-length": {
                "type": "string",
                "default": "79"
              },
              "etag": {
                "type": "string",
                "default": "W/\"4f-ufyO2sebbgnihjU82J2Z2j+pI48\""
              },
              "date": {
                "type": "string",
                "default": "Tue, 08 Dec 2020 19:59:15 GMT"
              },
              "connection": {
                "type": "string",
                "default": "close"
              },
              "Vary": {
                "type": "string",
                "default": "Accept-Encoding"
              }
            }
          }
        }
      }
    },
...
  },
...
}